Fix: Genoa file renders incorrectly with current color code
Details:
A typo in the DeviceGray code used an array of doubles instead of floats, causing incorrect
values to be returned by currenthsbcolor.

EXPECTED DIFFERENCES
Genoa file 061-01.ps should render correctly again.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8976 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/zcolor.c
@@ -1194,7 +1194,7 @@ static int grayrange(i_ctx_t * i_ctx_p, ref *space, float *ptr)
 static int graybasecolor(i_ctx_t * i_ctx_p, ref *space, int base, int *stage, int *cont, int *stack_depth)
 {
     os_ptr op = osp;
-    double Gray, RGB[3];
+    float Gray, RGB[3];
 
     *cont = 0;
     *stage = 0;
@@ -1205,7 +1205,7 @@ static int graybasecolor(i_ctx_t * i_ctx_p, ref *space, int base, int *stage, in
 	} else
 	    return_error(e_typecheck);
     } else
-	Gray = (double)op->value.intval;
+	Gray = (float)op->value.intval;
 
     if (Gray < 0 || Gray > 1)
 	return_error(e_rangecheck);
